requirements.txt
aiohttp==3.8.3
apscheduler==3.11.0
discord.py==2.4.0
python-dotenv==1.0.1
